change pit-token size with 32bytes

+++ daemon/mw-nfd/mw-nfd-global.hpp
@@ -137,9 +137,9 @@ namespace nfd {
 extern bool g_workerTimerTriggerList[DCN_MAX_WORKERS];
 
 	typedef struct st_msg {
-		uint8_t workerId;
+		uint64_t workerId;
         size_t hashValue;
-		uint8_t CanBePrefix;
+		uint64_t CanBePrefix;
 	} ST_PIT_TOKEN;
 
     using namespace moodycamel;
